Improving scalability of buildsystem
Although it's pretty simple right now. It's a lot better to have it so
that we don't need to keep on adding new files to the Makefile every
time there's a new separate c file.

It's also generally good practice to have `.PHONY` in Makefiles to
specify that the names of the steps like `all` or `clean` are not actual
source files. This keeps it somewhat more portable to systems with a more
POSIX style make.